+++ lib/sound_util/filter/gain.rb
@@ -8,19 +8,12 @@ module SoundUtil
       define_immutable_version :gain
 
       def gain!(factor)
-        info = format_info
         mutate_frames! do |_frame_idx, samples|
-          samples.map { |sample| scale_sample(sample, factor, info) }
+          samples.map { |sample| encode_value(sample_to_float(sample) * factor) }
         end
       end
 
       alias * gain
-
-      private
-
-      def scale_sample(sample, factor, info)
-        (sample * factor).round.clamp(info[:min], info[:max])
-      end
     end
   end
 end